DAY-023: Move modifiers and views to their own files
