+++ Days/ViewsAndModifiers/ViewsAndModifiers/ContentView.swift
@@ -25,57 +25,6 @@ struct ContentView: View {
     }
 }
 
-struct CapsuleText: View {
-    var text: String
-
-    var body: some View {
-        Text(text)
-            .font(.largeTitle)
-            .padding()
-            .background(.blue)
-            .clipShape(.capsule)
-    }
-}
-
-struct Title: ViewModifier {
-    func body(content: Content) -> some View {
-        content
-            .font(.largeTitle)
-            .foregroundStyle(.white)
-            .padding()
-            .background(.blue)
-            .clipShape(.rect(cornerRadius: 10))
-    }
-}
-
-extension View {
-    func titleStyle() -> some View {
-        modifier(Title())
-    }
-}
-
-struct Watermark: ViewModifier {
-    var text: String
-
-    func body(content: Content) -> some View {
-        ZStack(alignment: .bottomTrailing) {
-            content
-
-            Text(text)
-                .font(.caption)
-                .foregroundStyle(.white)
-                .padding(5)
-                .background(.black)
-        }
-    }
-}
-
-extension View {
-    func watermarked(with text: String) -> some View {
-        modifier(Watermark(text: text))
-    }
-}
-
 #Preview {
     ContentView()
 }
