+++ Days/ViewsAndModifiers/ViewsAndModifiers/Title.swift
@@ -1,0 +1,28 @@
+//
+// Title.swift
+// Last modified by 'Ángel Vázquez' in 2024.
+//
+
+import SwiftUI
+
+struct Title: ViewModifier {
+    func body(content: Content) -> some View {
+        content
+            .font(.largeTitle)
+            .foregroundStyle(.white)
+            .padding()
+            .background(.blue)
+            .clipShape(.rect(cornerRadius: 10))
+    }
+}
+
+extension View {
+    func titleStyle() -> some View {
+        modifier(Title())
+    }
+}
+
+#Preview {
+    Text("Hello world!")
+        .titleStyle()
+}
